--- server/clip_server/model/utils.py
@@ -1,1 +1,0 @@
-from open_clip.utils import to_2tuple
